refactor(ci): colocate error types in a single error module
RFC-0006 phase 2e follow-through (issue #450): the seven error enums
scattered across cuenv-ci (ExecutorError, RunnerError, CompilerError,
FlakeLockError, ValidationError, EmitterError, DiffError, GCError) move
verbatim into a new crates/ci/src/error.rs, together with the
From<ExecutorError> for cuenv_core::Error bridge, the EmitterResult
alias, and the FlakeLockError constructors and unit tests. The defining
modules re-export their type at the old path, so no consumer changes.

Deliberately no unified CiError mega-enum: each pipeline stage keeps
its dedicated error, no consumer matches across stages, and a catch-all
would recreate the god-error shape phase 2e removed from cuenv-core
(rationale recorded in the module doc).

Validation: cargo check -p cuenv-ci --all-targets clean; cargo test
-p cuenv-ci green (334 passed); cargo clippy -p cuenv-ci -p cuenv
--all-targets clean; cargo fmt; git diff --check.
